Move mock Airtable environment into test helpers file
